refactor: consolidate post-inference checks under validators
