+++ crates/semantics/src/lib.rs
@@ -4,10 +4,8 @@ pub mod call_classification;
 pub mod checker;
 pub mod context;
 pub mod facts;
-pub mod lint;
 pub mod loader;
 pub mod module_graph;
-pub mod pattern_analysis;
 pub mod prelude;
 pub mod store;
 pub mod validators;
