+++ crates/semantics/src/validators/mod.rs
@@ -1,35 +1,30 @@
-//! Post-inference validators.
-//!
-//! Each validator is an independent read-only walk over the frozen typed AST
-//! emitting diagnostics via a shared `DiagnosticSink`. Validators do not read
-//! the checker's `TypeEnv` (the freeze pass substitutes bound type variables
-//! before this stage runs) and do not share state with each other.
-//!
-//! Adding a validator: create a new submodule with a `pub(crate) fn run`
-//! taking `(&[Expression], &DiagnosticSink)` (plus `&Facts` if needed) and
-//! register it in `run_all` below.
-
 use diagnostics::DiagnosticSink;
 use syntax::ast::Expression;
-use syntax::program::CoercionInfo;
+use syntax::program::{CoercionInfo, UnusedInfo};
 
 use crate::context::AnalysisContext;
 use crate::facts::Facts;
 
+mod ast_lints;
 mod duplicate_bindings;
 mod enum_variant_value;
 mod generics;
 mod irrefutable_patterns;
+mod lints;
 mod native_value_usage;
 mod newtype;
+pub mod pattern_analysis;
 mod post_inference;
 mod prelude_shadowing;
 mod receivers;
+mod ref_lints;
 pub(crate) mod temp_producing;
 mod unused_expressions;
 mod visibility;
 
-pub struct ValidatorContext<'a> {
+pub use lints::Lint;
+
+pub(crate) struct ValidatorContext<'a> {
     pub typed_ast: &'a [Expression],
     pub is_typedef: bool,
     pub module_id: &'a str,
@@ -39,7 +34,50 @@ pub struct ValidatorContext<'a> {
     pub sink: &'a DiagnosticSink,
 }
 
-pub fn run_all(ctx: &mut ValidatorContext<'_>) {
+pub fn run(
+    analysis: &AnalysisContext,
+    facts: &mut Facts,
+    coercions: &CoercionInfo,
+    sink: &DiagnosticSink,
+    unused: &mut UnusedInfo,
+    run_lints: bool,
+) {
+    let store = analysis.store;
+
+    for module in store.modules.values() {
+        visibility::run_module(&module.id, store, sink);
+        for file in module.files.values() {
+            let is_typedef = module.typedefs.contains_key(&file.id);
+            let mut ctx = ValidatorContext {
+                typed_ast: &file.items,
+                is_typedef,
+                module_id: &module.id,
+                analysis,
+                facts,
+                coercions,
+                sink,
+            };
+            run_per_file(&mut ctx);
+        }
+    }
+
+    let pattern_ctx = pattern_analysis::Context::new(analysis, &facts.or_pattern_error_spans);
+    for module in store.modules.values() {
+        for file in module.files.values() {
+            for expression in &file.items {
+                pattern_analysis::check(expression, &pattern_ctx, sink);
+            }
+        }
+    }
+    facts.pattern_issues = pattern_ctx.take_issues();
+
+    if run_lints {
+        lints::lint_all_facts(facts, unused, sink);
+        lints::lint_all_modules(analysis, facts, sink, unused);
+    }
+}
+
+fn run_per_file(ctx: &mut ValidatorContext<'_>) {
     let store = ctx.analysis.store;
     duplicate_bindings::run(ctx.typed_ast, ctx.sink);
     irrefutable_patterns::run(ctx.typed_ast, ctx.sink);
@@ -58,6 +96,5 @@ pub fn run_all(ctx: &mut ValidatorContext<'_>) {
     enum_variant_value::run(ctx.typed_ast, store, ctx.sink);
     temp_producing::run(ctx.typed_ast, ctx.coercions, ctx.sink);
     unused_expressions::run(ctx.typed_ast, ctx.module_id, store, ctx.facts);
-    visibility::run_module(ctx.module_id, store, ctx.sink);
     post_inference::run(ctx.facts, ctx.sink);
 }
