+++ crates/semantics/src/validators/pattern_analysis/normalize.rs
