+++ crates/semantics/src/validators/pattern_analysis/types.rs
