+++ crates/semantics/src/validators/pattern_analysis/witness.rs
